+++ panos/data_source_system_info_test.go
@@ -1,0 +1,30 @@
+package panos
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform/helper/resource"
+)
+
+func TestPanosSystemInfo_basic(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccSystemConfig(),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet("data.panos_system_info.test", "version_major"),
+					resource.TestCheckResourceAttrSet("data.panos_system_info.test", "version_minor"),
+					resource.TestCheckResourceAttrSet("data.panos_system_info.test", "version_patch"),
+				),
+			},
+		},
+	})
+}
+
+func testAccSystemConfig() string {
+	return `
+data "panos_system_info" "test" {}
+`
+}